+++ metacopy/i18n/locales/en.ts
@@ -2,7 +2,7 @@ export default {
 	title: "MetaCopy Settings",
 	keyTitle: {
 		title: "Key",
-		desc: "The key which you want to copy the value",
+		desc: "The frontmatterKey which you want to copy the correspondingValue",
 		placeholder: "key1, key2, key3,…",
 	},
 	linkCreator: {
@@ -11,23 +11,23 @@ export default {
 		baseLinkDesc: "The base of the link",
 		behavior: {
 			title: "Default behavior",
-			desc: "Choose between a metadata key, obsidian path & fixed folder for the link creator",
+			desc: "Choose between a metadata frontmatterKey, obsidian path & fixed folder for the link creator",
 			fixedFolder: "Fixed Folder",
 			categoryKey: "Metadata Key",
 			obsidianPath: "Obsidian Path",
 		},
 		keyLink: {
 			title: "Key link",
-			desc: "The key to create as link",
-			defaultValue: "Default value",
-			defaultValueDesc: "If you want to active the link creation without the key set.",
+			desc: "The frontmatterKey to create as link",
+			defaultValue: "Default correspondingValue",
+			defaultValueDesc: "If you want to active the link creation without the frontmatterKey set.",
 		},
 		folderNote: {
 			title: "Folder note",
-			desc: "If the file name = key link or parent folder, remove the file name in the link"
+			desc: "If the file name = frontmatterKey link or parent folder, remove the file name in the link"
 		},
 		useFrontMatterTitle: {
-			title: "Use frontmatter key as title",
+			title: "Use frontmatter frontmatterKey as title",
 			desc: "Use a frontmatter field instead of the file name.",
 		},
 		regexReplaceTitle: {
@@ -41,14 +41,14 @@ export default {
 	},
 	disable: {
 		title: "Disable MetaCopy",
-		desc: "Disable Metacopy context menu with a frontmatter key.",
+		desc: "Disable Metacopy context menu with a frontmatter frontmatterKey.",
 		descURL: "Also disable the URL creation in command modal.",
 	},
 	menuBehavior: {
 		title: "Menu behavior",
-		desc: "Enable : require a configured key to enable the menu",
+		desc: "Enable : require a configured frontmatterKey to enable the menu",
 		keyMenu: "Key menu",
-		keyMenuDesc: "The key used to disable/enable the metacopy file menu",
+		keyMenuDesc: "The frontmatterKey used to disable/enable the metacopy file menu",
 	},
 
 	command : {